+++ examples/Tutorial/24_osc_to_supercollider/osc_to_supercollider.py
@@ -59,6 +59,6 @@ while vib.num_notes_playing() > 0:
     wait(1)
 
 performance = s.stop_transcribing()
-engraving_settings.max_voices_per_part = 1
+engraving_settings.max_voices_per_staff = 1
 performance.to_score(max_divisor=6, simplicity_preference=3).show()